compiler: fix object allocation (experimental)
When an object's member (a slot value) is dynamically
allocated, then it must itself be dynamically allocated
so that the garbage collector keeps the member alive.

I modified the compiler macro (dhm-t) for 'new-empty'
to have it trigger dynamic object allocation. But not
sure this is the best way to do it, hence 'experimental'.



git-svn-id: svn://svn.code.sf.net/p/lush/code@651 0babfeb2-de4b-461b-a2bf-1ab085750f0d